Our Creation Takes Form
Thirty-Six days of destruction. Four hours of creation. The pearly Logic
Gates open to the sweet hum of rejoicing. Let there be life. Let there
be sweet triangle set-up and fragment processing upon the lcd screen of
legend. Let the world look upon us, and give us peace when we are
finished.

So it begins.

Status: Motherfucking runnable